CollectionInterface creation and integration (partial) - #47
After a whole lot of refactoring and adding ALL of the new methods I wanted to add, I have finally gotten a SINGLE testcase class passing. So I'm committing it. This commit covers a LOT of ground though. I'll try to list all of it but I'm sure I'll miss something:

* Added indexOf() method - refs #20
* Added random() method - refs #39
* Added namespaced helper functions collect(), invoke(), is_traversable(), is_arrayable(), to_array(), and typeof() - refs #42
* Renamed getValueAtPosition() and getKeyAtPosition() to getOffset() and getOffsetKey() - refs #45
* Added CollectionInterface - refs #39
* Added Arrayable support with ArrayableInterface and is_arrayable() - refs #23
* Using Laravel's collections for inspiration, I rounded out my main collection mehods which are defined in CollectionInterface - refs #30
* Added to_array() method which is somewhat similar to Laravel's getArrayableItems(), but it still needs some work - refs #31
* A whole bunch of other minor changes

I still need to get the remaining unit tests working and I need to add unit tests for the new methods I added with this commit.