Added immutability tests - refs #47
Added some methods to my base unit test case class that allows me to "watch" an object over the course of a test and then at the end, assert that it has not changed in any discernable way. This will help me immensely while trying to unit test that my collection methods are respecting collection object immutability. I have implemented this assert method in all existing tests that warrant it.